backend/requirements.txt
fastapi
uvicorn
supabase
pydantic-settings
python-jose
